refactor(tui::ui::components::popups::saveplaylist): move function "*mount_save_playlist*" to popup file
