+++ tui/src/ui/components/popups/mod.rs
@@ -23,4 +23,5 @@ pub use podcast::{
 };
 #[allow(unused_imports)]
 pub use quit::QuitPopup;
+#[allow(unused_imports)]
 pub use saveplaylist::{SavePlaylistConfirmPopup, SavePlaylistPopup};
